added pagination

+++ internal/business/domains/v1/domain.petitions.go
@@ -10,8 +10,8 @@ type PetitionDomain struct {
 	Title       string
 	Description string
 	OwnerID     string
-	Likes       int64
-	Voices      int64
+	Likes       int
+	Voices      int
 	CreatedAt   time.Time
 	UpdatedAt   time.Time
 }
@@ -22,7 +22,7 @@ type PetitionUseсase interface {
 	Delete(ctx context.Context, id string, userId string, userRoles []string) (err error)
 	Like(ctx context.Context, id string, userId string) (err error)
 	Voice(ctx context.Context, id string, userId string) (err error)
-	GetAll(ctx context.Context, pageNumber int64, pageSize int64) (outDomains []*PetitionDomain, total int64, err error)
+	GetAll(ctx context.Context, pageNumber int, pageSize int) (outDomains []*PetitionDomain, total int, err error)
 }
 
 type PetitionRepository interface {
@@ -31,6 +31,6 @@ type PetitionRepository interface {
 	Delete(ctx context.Context, id string) (err error)
 	Like(ctx context.Context, id string, userId string) (err error)
 	Voice(ctx context.Context, id string, userId string) (err error)
-	GetAll(ctx context.Context, pageNumber int64, pageSize int64) (outDomains []*PetitionDomain, total int64, err error)
+	GetAll(ctx context.Context, pageNumber int, pageSize int) (outDomains []*PetitionDomain, total int, err error)
 	GetByID(ctx context.Context, id string) (outDomain *PetitionDomain, err error)
 }
