+++ internal/http/datatransfers/responses/v1/response.base.go
@@ -2,7 +2,7 @@ package v1
 
 type PageResponse struct {
 	Data        interface{} `json:"data"`
-	CurrentPage int64       `json:"current_page"`
-	PageSize    int64       `json:"page_size"`
-	TotalPages  int64       `json:"total_pages"`
+	CurrentPage int         `json:"current_page"`
+	PageSize    int         `json:"page_size"`
+	TotalPages  int         `json:"total_pages"`
 }
